parse-man: fix bug in description formatting...
with a hack:
program listing should be indented dirently. But the cleanup of 'para'
element removes the indentation present in xml file. This also removes
the indentation of the programlisting that should be added to respect
pod format. Hence the special hacking tag '+-+' is added at cleanup
time. Later this tag is replaced with a proper indentation while
cleaning up the element description.